% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_heap.R
\docType{class}
\name{binomial_heap-class}
\alias{binomial_heap-class}
\title{Binomial heap class}
\description{
Implementation of a binomial heap datastructure, i.e. a
 priority datastructure with \code{push} and \code{pop} in amortized
 \emph{O(log n)}. \code{binomial_heap} wraps a Boost
 \code{binomial_heap} using Rcpp modules. The heap consists of nodes with
 keys and values where the key determines the priority in the heap.
}
\section{Slots}{

\describe{
\item{\code{.data}}{object that bundles all important heap related objects}
}}

