% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap_binomial.R
\name{binomial_heap}
\alias{binomial_heap}
\title{Create a new \code{binomial_heap}}
\usage{
binomial_heap(key.class = c("character", "numeric", "integer"),
  value.class = c("character", "numeric", "integer"))
}
\arguments{
\item{key.class}{the primitive class type of the keys}

\item{value.class}{the primitive class type of the values}
}
\value{
returns a new \code{binomial_heap} object
}
\description{
Instantiates a new \code{\linkS4class{binomial_heap}} object,
 i.e. a tree-like data structure satisfying the \emph{min-heap} property.
}
\examples{
b <- binomial_heap("numeric", "numeric")
b[1] <- rnorm(5)
b <- insert(b, c(2, 3), matrix(rnorm(10), 2))
b <- insert(b, c(4, 5), list(rnorm(2), 4))
b[c(-1, 7)] <- rnorm(2)
pop(b)
size(b)
peek(b)

b <- binomial_heap("character", "numeric")
b[letters] <- rnorm(26)
pop(b)
b["?"] <- 2
peek(b)

}
