% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_clear.R, R/ds_deque.R, R/ds_heap.R,
%   R/ds_map.R
\docType{methods}
\name{clear}
\alias{clear}
\alias{clear,deque-method}
\alias{clear,heap-method}
\alias{clear,map-method}
\title{Remove all elements from a datastructure}
\usage{
clear(obj)

\S4method{clear}{deque}(obj)

\S4method{clear}{heap}(obj)

\S4method{clear}{map}(obj)
}
\arguments{
\item{obj}{the object to clear}
}
\description{
Removes every element that is stored in a data structure and
 resets everything.
}
