% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastructures-package.R
\docType{package}
\name{datastructures-package}
\alias{datastructures-package}
\title{datastructures}
\description{
Implementation of  advanced datastructures such as hashmaps,
heaps, or queues . Advanced datastructures are essential
in many computer science and statistics problems such as for graph
algorithms or string analysis. datastructures use Boost and Rcpp modules
and wrap these as C++ objects for R usage.
}
\author{
Simon Dirmeier
}
\keyword{package}
