% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_deque.R
\docType{class}
\name{deque-class}
\alias{deque-class}
\title{Deque class}
\description{
Abstract deque class
}
\section{Slots}{

\describe{
\item{\code{.deque}}{\code{C++} object representing a deque}

\item{\code{.key.class}}{the class of the keys}
}}

