% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_erase.R, R/ds_map.R, R/ds_map_bimap.R,
%   R/ds_map_multimap.R
\docType{methods}
\name{erase}
\alias{erase}
\alias{erase,map,vector,missing-method}
\alias{erase,bimap,missing,vector-method}
\alias{erase,multimap,vector,vector-method}
\alias{erase,multimap,vector,list-method}
\alias{erase,multimap,vector,ANY-method}
\title{Erase an entry from a map}
\usage{
erase(obj, key, value)

\S4method{erase}{map,vector,missing}(obj, key)

\S4method{erase}{bimap,missing,vector}(obj, value)

\S4method{erase}{multimap,vector,vector}(obj, key, value)

\S4method{erase}{multimap,vector,list}(obj, key, value)

\S4method{erase}{multimap,vector,ANY}(obj, key, value)
}
\arguments{
\item{obj}{the object to pop an element from}

\item{key}{a vector of keys that should be removed}

\item{value}{optionally a list of values needs to be supplied for some data
structures such as multimaps if entries should be removed.}
}
\value{
returns \code{obj} with removed values
}
\description{
Erase a vector of key-value pair from a map object.
}
