% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_fibonacci_heap.R
\docType{class}
\name{fibonacci_heap-class}
\alias{fibonacci_heap-class}
\title{Fibonacci heap class}
\description{
Implementation of a Fibonacci heap datastructure, i.e. a
 priority datastructure with \code{push} in amortized \emph{O(1)} and \code{pop}
 in \emph{O(log n)}. \code{fibonacci_heap} wraps a Boost \code{fibonacci_heap}
 using Rcpp modules. The heap consists of nodes with keys and values where the key
 determines the priority in the heap.
}
\section{Slots}{

\describe{
\item{\code{.data}}{object that bundles all important heap related objects}
}}

\examples{
 # a fibonacci_heap with <character, double> key-value nodes
 fibonacci_heap <- new("fibonacci_heap", "character", "numeric")

 # a fibonacci_heap with <character, integer> key-value nodes
 fibonacci_heap <- new("fibonacci_heap", "character", "integer")

 # a fibonacci_heap with <double, double> key-value nodes
 fibonacci_heap <- new("fibonacci_heap", "numeric", "numeric")

 # a fibonacci_heap with <integer, double> key-value nodes
 fibonacci_heap <- new("fibonacci_heap", "integer", "numeric")

 # a fibonacci_heap with <integer, character> key-value nodes
 fibonacci_heap <- new("fibonacci_heap", "integer", "character")
}
