% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{methods}
\name{g}
\alias{g}
\alias{g,matrix-method}
\title{Do another thing}
\usage{
g(X, Y)

\S4method{g}{matrix}(X, Y)
}
\arguments{
\item{X}{some object}

\item{Y}{some other object}
}
\description{
TODO
}
