% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_get.R
\docType{methods}
\name{get}
\alias{get}
\alias{get,hashmap-method}
\title{Get elements from an object}
\usage{
get(obj, x)

\S4method{get}{hashmap}(obj, x)
}
\arguments{
\item{obj}{object to extract values from}

\item{x}{the set of keys to match the values}
}
\value{
returns extracted values from the object
}
\description{
Extracts a set of <key, value> pairs from an object using a
 set of keys.
}
