% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_hashmap.R
\name{hashmap}
\alias{hashmap}
\title{Create a new \code{hashmap}}
\usage{
hashmap(key.class = c("character", "numeric", "integer"),
  value.class = c("character", "numeric", "integer"))
}
\arguments{
\item{key.class}{the primitive class type of the keys}

\item{value.class}{the primitive class type of the values}
}
\value{
returns a new \code{hashmap} object
}
\description{
Instantiates a new \code{\linkS4class{hashmap}} object,
 i.e. an unordered collection of key-value pairs with mapping
 \deqn{f: keys -> values.}
}
