% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_head.R
\docType{methods}
\name{head}
\alias{head}
\title{Have a look at the head of an object}
\usage{
head(x, ...)
}
\arguments{
\item{x}{the object to peek}

\item{...}{other arguments that are only required for \code{utils::head}}
}
\value{
returns the head of an object
}
\description{
Shows the first few elements of a data structure.
}
\examples{

 # shows the head of a hashmap
 h_map <- hashmap()
 h_map[letters] <- rnorm(length(letters))
 head(h_map)

 # shows the head of a bimap
 b_map <- bimap("integer", "integer")
 b_map[seq(10)] <- seq(10, 1)
 head(b_map)

}
