% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,hashmap-method}
\title{Add elements to an object}
\usage{
insert(obj, x, y)

\S4method{insert}{hashmap}(obj, x, y)
}
\arguments{
\item{obj}{the object you want to insert elements to}

\item{x}{a scalar/vector of keys}

\item{y}{a scalar/vector of values}
}
\examples{
 # insert to a hashmap
 hashmap <- new("hashmap", "character", "numeric")
 hashmap <- insert(hashmap, "test", 1.)
 hashmap <- insert(hashmap, paste0(1:10), 1:10)
}
