% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map.R
\docType{class}
\name{map-class}
\alias{map-class}
\title{Map class}
\description{
Abstract map class
}
\section{Slots}{

\describe{
\item{\code{.map}}{\code{C++} object representing a mapping}

\item{\code{.key.class}}{the class of the keys}

\item{\code{.value.class}}{the class of the values}
}}

