% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_multimap.R
\name{multimap}
\alias{multimap}
\title{Create a new \code{multimap}}
\usage{
multimap(key.class = c("character", "numeric", "integer"),
  value.class = c("character", "numeric", "integer"))
}
\arguments{
\item{key.class}{the primitive class type of the keys}

\item{value.class}{the primitive class type of the values}
}
\value{
returns a new \code{multimap} object
}
\description{
Instantiates a new \code{\linkS4class{multimap}} object,
 i.e. an unordered collection of key-value pairs with mapping
 \deqn{f: keys -> values}, where multiple identical key-value paors
 can be stored.
}
