% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_queue.R
\docType{class}
\name{queue-class}
\alias{queue-class}
\title{Queue class}
\description{
Implementation of a queue datastructure, i.e. a list
 implementation with FIFO principle. \code{queue} uses a \code{std::deque}
 as default container, so inserting, peeking and popping functions require
 constant \emph{O(1)}.
}
\section{Slots}{

\describe{
\item{\code{.data}}{object that bundles all important heap related objects}
}}

