% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_deque_queue.R
\name{queue}
\alias{queue}
\title{Create a new \code{queue}}
\usage{
queue(key.class = c("character", "numeric", "integer"))
}
\arguments{
\item{key.class}{the primitive class type of the keys}
}
\value{
returns a new \code{queue} object
}
\description{
Instantiates a new \code{\linkS4class{queue}} object,
 i.e. a list implementation with FIFO principle.
}
\examples{
 q <- queue("numeric")
 q <- insert(q, rnorm(10))
 q <- insert(q, 2)
 q <- insert(q, as.list(rnorm(5)))
 peek(q)
 size(q)

 q <- queue("character")
 q <- insert(q, letters)
 q <- insert(q, "hello")
 q <- insert(q, as.list(letters))
 pop(q)
 size(q)

}
