% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_size.R, R/ds_deque_queue.R,
%   R/ds_deque_stack.R, R/ds_heap_binomial.R, R/ds_heap_fibonacci.R,
%   R/ds_map_bimap.R, R/ds_map_hashmap.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,queue-method}
\alias{size,stack-method}
\alias{size,binomial_heap-method}
\alias{size,fibonacci_heap-method}
\alias{size,bimap-method}
\alias{size,hashmap-method}
\title{Get the size of an object}
\usage{
size(obj)

\S4method{size}{queue}(obj)

\S4method{size}{stack}(obj)

\S4method{size}{binomial_heap}(obj)

\S4method{size}{fibonacci_heap}(obj)

\S4method{size}{bimap}(obj)

\S4method{size}{hashmap}(obj)
}
\arguments{
\item{obj}{the object to get the size from}
}
\value{
returns the size of \code{obj}
}
\description{
Computes the size of an object, i.e. the number of keys or
 <key, value> pairs depending on the object.
}
