% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_stack.R
\docType{class}
\name{stack-class}
\alias{stack-class}
\title{Stack class}
\description{
Implementation of a stack datastructure, i.e. a list
 implementation with LIFO principle. \code{stack} uses a \code{std::deque}
 as default container, so inserting, peeking and popping functions require
 constant \emph{O(1)}.
}
\section{Slots}{

\describe{
\item{\code{.data}}{object that bundles all important heap related objects}
}}

\examples{
 # a stack with <character> elements
 stack <- new("stack", "character")

 # a stack with <numeric> elements
 stack <- new("stack", "numeric")

 # a stack with <integer> elements
 stack <- new("stack", "integer")

}
