% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_hashmap.R
\docType{methods}
\name{[,hashmap,ANY,missing,missing-method}
\alias{[,hashmap,ANY,missing,missing-method}
\title{Extract parts to an object}
\usage{
\S4method{[}{hashmap,ANY,missing,missing}(x, i, j = "missing", ...,
  drop = "missing")
}
\description{
Access <key, value> pairs of a hashmap.

 @param x  a \code{hashmap}
 @param i  a vector of keys
}
