% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_unordered.R
\docType{methods}
\name{[,unordered_map,vector,missing,missing-method}
\alias{[,unordered_map,vector,missing,missing-method}
\title{Extract elements from an object}
\usage{
\S4method{[}{unordered_map,vector,missing,missing}(x, i)
}
\arguments{
\item{x}{a \code{hashmap}}

\item{i}{a vector of keys}
}
\description{
Access <key, value> pairs of an unordered map using a set of keys.
}
