% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_insert.R
\docType{methods}
\name{[<-,binomial_heap,ANY,missing-method}
\alias{[<-,binomial_heap,ANY,missing-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{binomial_heap,ANY,missing}(x, i) <- value
}
\arguments{
\item{x}{a \code{heap}}

\item{i}{a vector of keys}

\item{value}{a vector of values for the keys}
}
\description{
Inserts <key, value> pairs to a binomial heap. The keys are
 determine the ordering of the heap, while the value is the actual value to
 store.
}
