% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_fibonacci_heap.R
\docType{methods}
\name{[<-,fibonacci_heap,ANY,missing-method}
\alias{[<-,fibonacci_heap,ANY,missing-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{fibonacci_heap,ANY,missing}(x, i, j = "missing", ...) <- value
}
\description{
Inserts <key, value> pairs to a Fibonacci heap. The keys are
 determine the ordering of the heap, while the value is the actual value to
 store.

 @param x  a \code{fionacci_heap}
 @param i  a vector of keys
 @param value  a vector of values for the keys
}
