% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap_fibonacci.R
\docType{methods}
\name{[<-,fibonacci_heap,vector,missing,list-method}
\alias{[<-,fibonacci_heap,vector,missing,list-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{fibonacci_heap,vector,missing,list}(x, i) <- value
}
\arguments{
\item{x}{a \code{heap}}

\item{i}{a vector of keys}

\item{value}{a vector of values for the keys}
}
\description{
Inserts <key, value> pairs to a Fibonacci heap. The keys are
 determine the ordering of the heap, while the value is the actual value to
 store.
}
