% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_hashmap.R
\docType{methods}
\name{[<-,hashmap,ANY,missing-method}
\alias{[<-,hashmap,ANY,missing-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{hashmap,ANY,missing}(x, i, j = "missing", ...) <- value
}
\description{
Inserts <key, value> pairs to a hashmap.

 @param x  a \code{hashmap}
 @param i  a vector of keys
 @param value  a vector of values for the keys
}
