% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_insert.R
\docType{methods}
\name{[<-,map,ANY,missing-method}
\alias{[<-,map,ANY,missing-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{map,ANY,missing}(x, i) <- value
}
\arguments{
\item{x}{a \code{map} object}

\item{i}{a vector of keys}

\item{value}{a vector of values for the keys}
}
\description{
Inserts <key, value> pairs to a map.
}
