% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_unordered.R
\docType{methods}
\name{[<-,unordered_map,vector,missing,matrix-method}
\alias{[<-,unordered_map,vector,missing,matrix-method}
\title{Insert parts to an object}
\usage{
\S4method{[}{unordered_map,vector,missing,matrix}(x, i) <- value
}
\arguments{
\item{x}{a \code{map} object}

\item{i}{a vector of keys}

\item{value}{a vector of values for the keys}
}
\description{
Inserts <key, value> pairs to an unordered map.
}
