% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_unordered.R
\docType{class}
\name{unordered_map-class}
\alias{unordered_map-class}
\title{Abstract unordered map class}
\description{
Abstract unordered map class
}
\section{Slots}{

\describe{
\item{\code{.map}}{\code{C++} object representing a mapping}

\item{\code{.key.class}}{the class of the keys}
}}

