% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/vector.R
\name{as.list.Vector}
\alias{as.list.Vector}
\alias{vector_create}
\alias{vector_empty}
\alias{vector_size}
\alias{vector_front}
\alias{vector_back}
\alias{vector_push_back}
\alias{vector_pop_back}
\alias{vector_at}
\alias{vector_set_at}
\alias{Vector}
\alias{print.Vector}
\alias{format.Vector}
\alias{[.Vector}
\alias{[<-.Vector}
\alias{vector_at<-}
\title{Vector}
\usage{
\method{as.list}{Vector}(vec)

vector_create(n = 0L)

vector_empty(vec)

vector_size(vec)

vector_front(vec)

vector_back(vec)

vector_push_back(vec, obj)

vector_pop_back(vec)

vector_at(vec, i)

vector_set_at(vec, i, obj)

\method{print}{Vector}(vec, ...)

\method{format}{Vector}(vec, ...)

\method{[}{Vector}(vec, i)

\method{[}{Vector}(vec, i) <- value

vector_at(vec, i) <- value
}
\arguments{
\item{vec}{a vector object}

\item{n}{vector size}

\item{obj}{an R object}

\item{i}{position in the vector}

\item{...}{other arguments passed to or from other methods}

\item{value}{object to set}
}
\value{
\code{as.list.Vector} returns an R list object.

\code{vector_create} returns a new, empty vector.

\code{vector_empty} returns a single logical value.

\code{vector_size} returns a single integer value.

\code{vector_front} returns a single RObject value or
throws an error if a vector is empty.

\code{vector_back} returns a single RObject value or
throws an error if a vector is empty.

\code{vector_push_back} does not return anything interesting.

\code{vector_pop_back} does not return anything interesting. Throws an
error if a vector is empty.

\code{vector_at} returns a single RObject value.

\code{vector_set_at} does not return anything interesting.

\code{print.Vector} returns the vector coerced to a list, invisibly.

\code{format.Vector} returns the character representation
of the objects in the vector.

\code{[.Vector} returns a single R object.

\code{[<-.Vector} returns a modified vector

\code{vector_at<-} returns a modified vector
}
\description{
Vectors are sequence containers representing arrays that can change in size.
}
\details{
\code{as.list.Vector} converts a given vector object to an R list.

\code{vector_create} creates a new vector.

\code{vector_empty} determines if a given stack is empty or not.

\code{vector_size} returns a size of vector.

\code{vector_front} returns the first object of a given vector.

\code{vector_back} returns the last object of a given vector.

\code{vector_push_back} pushes a given object at the end of the vector.

\code{vector_pop_back} removes the last element in the vector, effectively
reducing the container size by one.

\code{vector_at} returns an object at position i in the vector.

\code{vector_set_at} sets an object obj at position i in the vector.

\code{print.Vector} prints the contents of the vector
(coerced to a list object) on the console.

\code{format.Vector} pretty-prints the contents of the vector
(coerced to a list object) on the console.

\code{[.Vector} returns an object at position i in the vector.

\code{[<-.Vector} sets an object value at position i in the vector.

\code{vector_at<-} sets an object value at position i in the vector.
}
