% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap_binomial.R
\docType{class}
\name{binomial_heap-class}
\alias{binomial_heap-class}
\title{Binomial heap class}
\description{
Implementation of a binomial heap data structure, i.e. a
 priority datastructure with \code{push} and \code{pop} in amortized
 \emph{O(log n)}. \code{binomial_heap} wraps a \code{boost::binomial_heap}
 using Rcpp modules. The heap consists of nodes with
 keys and values where the key determines the priority in the heap.
 Also see the \code{\linkS4class{binomial_heap}} class.
}
\section{Slots}{

\describe{
\item{\code{.heap}}{\code{C++} object representing a heap}

\item{\code{.key.class}}{the class of the keys}
}}

