% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_clear.R
\docType{methods}
\name{clear}
\alias{clear}
\title{Remove all elements from a datastructure}
\usage{
clear(obj)
}
\arguments{
\item{obj}{the object to clear}
}
\description{
Removes every element that is stored in a data structure and
 resets everything.
}
