% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastructures-package.R
\docType{package}
\name{datastructures-package}
\alias{datastructures-package}
\title{datastructures}
\description{
Implementation of core datastructures for R.
}
\author{
Simon Dirmeier
}
\keyword{package}
