% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap_fibonacci.R
\docType{class}
\name{fibonacci_heap-class}
\alias{fibonacci_heap-class}
\title{Fibonacci heap class}
\description{
Implementation of a Fibonacci heap data structure, i.e. a
 priority datastructure with \code{push} in amortized \emph{O(1)} and
 \code{pop} in \emph{O(log n)}. \code{fibonacci_heap} wraps a
 \code{boost::fibonacci_heap} using Rcpp modules. The heap consists of
 nodes with keys and values where the key determines the priority in the
 heap. Also see the \code{\linkS4class{binomial_heap}} class.
}
\section{Slots}{

\describe{
\item{\code{.heap}}{\code{C++} object representing a heap}

\item{\code{.key.class}}{the class of the keys}
}}

