% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap_fibonacci.R
\name{fibonacci_heap}
\alias{fibonacci_heap}
\title{Create a new \code{fibonacci_heap}}
\usage{
fibonacci_heap(key.class = c("character", "numeric", "integer"),
  value.class = c("character", "numeric", "integer"))
}
\arguments{
\item{key.class}{the primitive class type of the keys}

\item{value.class}{the primitive class type of the values}
}
\value{
returns a new \code{fibonacci_heap} object
}
\description{
Instantiates a new \code{\linkS4class{fibonacci_heap}} object,
 i.e. a tree-like data structure satisfying the \emph{min-heap} property.
}
