% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_get.R, R/ds_hashmap.R
\docType{methods}
\name{get}
\alias{get}
\alias{get,hashmap-method}
\title{Get elements from an object}
\usage{
get(obj, x)

\S4method{get}{hashmap}(obj, x)
}
\arguments{
\item{obj}{object to extract values from}

\item{x}{the set of keys to match the values}
}
\value{
returns extracted values from the object
}
\description{
Extracts a set of <key, value> pairs from an object using a
 set of keys.
}
\examples{
\dontrun{
 # insert to a hashmap with <character, character> pairs
 hashmap <- new("hashmap", "character", "character")
 hashmap <- insert(hashmap, paste0("k", 1:10), paste0("v", 1:10))

 get(hashmap, paste0("k", c(3,5,7)))

 hashmap[paste0("k", c(3,5,7))]
}
}
