% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{get}
\alias{get}
\alias{get,hashmap-method}
\title{Get elements from an object}
\usage{
get(obj, x)

\S4method{get}{hashmap}(obj, x)
}
\arguments{
\item{obj}{the object you want to insert elements to}

\item{x}{a scalar/vector of keys}
}
\examples{
 # insert to a hashmap with <character, double> pairs
 hashmap <- new("hashmap", "character", "character")
 hashmap <- insert(hashmap, paste0("k", 1:10), paste0("v", 1:10))
 get(hashmap, paste0("k", c(3,5,7)))
}
