% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_hashmap.R
\docType{class}
\name{hashmap-class}
\alias{hashmap-class}
\title{Hashmap class}
\description{
Implementation of a hashmap data structure, i.e. an unordered
collection of key-value pairs:
\deqn{f: keys -> values.}
Hashmaps only to store unique keys-value pairs. For a data structure
where multiple identical keys can be stores see \code{\link{multimap}}.
Inserting and accessing is amortized in \emph{O(1)}.
\code{hashmap} wraps a C++ \code{unordered_map} using Rcpp modules.
Also see \code{\linkS4class{bimap}} for mappings in both ways.
}
\section{Slots}{

\describe{
\item{\code{.map}}{\code{C++} object representing a mapping}

\item{\code{.key.class}}{the class of the keys}
}}

\examples{
 # creates a hashmap<character, SEXP>
 h <- hashmap()

}
