% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_hashmap.R
\docType{class}
\name{hashmap-class}
\alias{hashmap-class}
\title{Hashmap class}
\description{
Implementation of a hashmap datastructure, i.e. an unordered collection
 of key-value pairs. Inserting and accessing is amortized in \emph{O(1)}.
 \code{hashmap} wraps a C++ \code{unordered_map} using Rcpp modules.
}
\section{Slots}{

\describe{
\item{\code{.data}}{object that bundles all important map related objects}
}}

\examples{
 # a hashmap with <character, double> pairs
 hashmap <- new("hashmap", "character", "numeric")

 # a hashmap with <character, integer> pairs
 hashmap <- new("hashmap", "character", "integer")

 # a hashmap with <numeric, numeric> pairs
 hashmap <- new("hashmap", "numeric", "numeric")
}
