% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\docType{class}
\name{hashmap-class}
\alias{hashmap-class}
\title{Hashmap class}
\description{
Implementation of a hashmap datastructure. Inserting and accesing
 is amortized \emph{O(1)}. The class wraps a C++ \code{unordered_map}.
}
\section{Slots}{

\describe{
\item{\code{.map}}{an Rcpp module wrapping an \code{unordered_map}}
}}

