% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_heap.R
\docType{class}
\name{heap-class}
\alias{heap-class}
\title{Abstract heap class}
\description{
Abstract heap class that bundles
}
\section{Slots}{

\describe{
\item{\code{.heap}}{\code{C++} object representing a heap}

\item{\code{.key.class}}{the class of the keys}

\item{\code{.value.class}}{the class of the values}
}}

