% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_insert.R, R/ds_deque_queue.R,
%   R/ds_deque_stack.R, R/ds_heap_binomial.R, R/ds_heap_fibonacci.R,
%   R/ds_map_bimap.R, R/ds_map_hashmap.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,queue,vector,missing-method}
\alias{insert,queue,list,missing-method}
\alias{insert,queue,matrix,missing-method}
\alias{insert,stack,vector,missing-method}
\alias{insert,stack,list,missing-method}
\alias{insert,stack,matrix,missing-method}
\alias{insert,binomial_heap,vector,vector-method}
\alias{insert,binomial_heap,vector,list-method}
\alias{insert,binomial_heap,vector,matrix-method}
\alias{insert,fibonacci_heap,vector,vector-method}
\alias{insert,fibonacci_heap,vector,list-method}
\alias{insert,fibonacci_heap,vector,matrix-method}
\alias{insert,bimap,vector,vector-method}
\alias{insert,hashmap,vector,vector-method}
\alias{insert,hashmap,vector,list-method}
\alias{insert,hashmap,vector,matrix-method}
\title{Add elements to an object}
\usage{
insert(obj, x, y)

\S4method{insert}{queue,vector,missing}(obj, x)

\S4method{insert}{queue,list,missing}(obj, x)

\S4method{insert}{queue,matrix,missing}(obj, x)

\S4method{insert}{stack,vector,missing}(obj, x)

\S4method{insert}{stack,list,missing}(obj, x)

\S4method{insert}{stack,matrix,missing}(obj, x)

\S4method{insert}{binomial_heap,vector,vector}(obj, x, y)

\S4method{insert}{binomial_heap,vector,list}(obj, x, y)

\S4method{insert}{binomial_heap,vector,matrix}(obj, x, y)

\S4method{insert}{fibonacci_heap,vector,vector}(obj, x, y)

\S4method{insert}{fibonacci_heap,vector,list}(obj, x, y)

\S4method{insert}{fibonacci_heap,vector,matrix}(obj, x, y)

\S4method{insert}{bimap,vector,vector}(obj, x, y)

\S4method{insert}{hashmap,vector,vector}(obj, x, y)

\S4method{insert}{hashmap,vector,list}(obj, x, y)

\S4method{insert}{hashmap,vector,matrix}(obj, x, y)
}
\arguments{
\item{obj}{object to insert into}

\item{x}{the values/keys to insert into}

\item{y}{values to be inserted which are required for some datastructures}
}
\value{
returns \code{obj} with inserted values
}
\description{
Adds keys or <key, value> pairs to an object and returns the
 object. Depending on the datastructure used, either only keys are required
 or pairs of <keys, values>.
}
