% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_insert.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,heap,ANY,ANY-method}
\alias{insert,map,ANY,ANY-method}
\alias{insert,deque,ANY,missing-method}
\title{Add elements to an object}
\usage{
insert(obj, x, y)

\S4method{insert}{heap,ANY,ANY}(obj, x, y)

\S4method{insert}{map,ANY,ANY}(obj, x, y)

\S4method{insert}{deque,ANY,missing}(obj, x)
}
\arguments{
\item{obj}{object to insert into}

\item{x}{the values/keys to insert into}

\item{y}{values to be inserted which are required for some datastructures}
}
\value{
returns \code{obj} with inserted values
}
\description{
Adds keys or <key, value> pairs to an object and returns the
 object. Depending on the datastructure used, either only keys are required
 or pairs of <keys, values>.
}
