% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,hashmap,character,character-method}
\alias{insert,character,character-method}
\title{Add key-value pairs to a \code{hashmap}-object}
\usage{
insert(obj, keys, values)

\S4method{insert}{hashmap,character,character}(obj, keys, values)
}
\arguments{
\item{obj}{the hashmap you want to insert pairs to}

\item{keys}{a scalar/vector of keys}

\item{values}{a scalar/vector of values}
}
\examples{
  hashmap <- new("hashmap")
  insert(hashmap, "test", 1)
}
