% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_insert.R, R/ds_deque.R, R/ds_heap.R,
%   R/ds_map_bimap.R, R/ds_map_hashmap.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,deque,vector,missing-method}
\alias{insert,deque,list,missing-method}
\alias{insert,deque,matrix,missing-method}
\alias{insert,heap,vector,vector-method}
\alias{insert,heap,vector,list-method}
\alias{insert,heap,vector,matrix-method}
\alias{insert,bimap,vector,vector-method}
\alias{insert,hashmap,vector,vector-method}
\alias{insert,hashmap,vector,list-method}
\alias{insert,hashmap,vector,matrix-method}
\title{Add elements to an object}
\usage{
insert(obj, x, y)

\S4method{insert}{deque,vector,missing}(obj, x)

\S4method{insert}{deque,list,missing}(obj, x)

\S4method{insert}{deque,matrix,missing}(obj, x)

\S4method{insert}{heap,vector,vector}(obj, x, y)

\S4method{insert}{heap,vector,list}(obj, x, y)

\S4method{insert}{heap,vector,matrix}(obj, x, y)

\S4method{insert}{bimap,vector,vector}(obj, x, y)

\S4method{insert}{hashmap,vector,vector}(obj, x, y)

\S4method{insert}{hashmap,vector,list}(obj, x, y)

\S4method{insert}{hashmap,vector,matrix}(obj, x, y)
}
\arguments{
\item{obj}{object to insert into}

\item{x}{the values/keys to insert into}

\item{y}{values to be inserted which are required for some datastructures}
}
\value{
returns \code{obj} with inserted values
}
\description{
Adds keys or <key, value> pairs to an object and returns the
 object. Depending on the datastructure used, either only keys are required
 or pairs of <keys, values>.
}
