% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R, R/fibonacci_heap.R, R/hashmap.R
\docType{methods}
\name{insert}
\alias{insert}
\alias{insert,fibonacci_heap-method}
\alias{insert,hashmap-method}
\title{Add elements to an object}
\usage{
insert(obj, x, y)

\S4method{insert}{fibonacci_heap}(obj, x, y)

\S4method{insert}{hashmap}(obj, x, y)
}
\arguments{
\item{obj}{the object you want to insert elements to}

\item{x}{a scalar/vector of keys}

\item{y}{a scalar/vector of values}

\item{obj}{the object you want to insert elements to}

\item{x}{a scalar/vector of keys}

\item{y}{a scalar/vector of values}
}
\examples{
 # insert to a fibonacci_heap with <character, double> nodes
 hashmap <- new("fibonacci_heap", "numeric", "double")

 hashmap <- insert(fibonacci_heap, 1.0, "test")

 hashmap[rnorm(5)] <- 1:5

 # insert to a hashmap with <character, double> pairs
 hashmap <- new("hashmap", "character", "integer")

 hashmap <- insert(hashmap, "test", 1)

 hashmap[paste0("k", 1:10)] <- 1:10

}
