% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_keys.R, R/ds_map_bimap.R,
%   R/ds_map_hashmap.R
\docType{methods}
\name{keys}
\alias{keys}
\alias{keys,bimap-method}
\alias{keys,hashmap-method}
\title{Get keys from an object}
\usage{
keys(obj)

\S4method{keys}{bimap}(obj)

\S4method{keys}{hashmap}(obj)
}
\arguments{
\item{obj}{object to extract keys from}
}
\value{
returns the extracted keys as vector
}
\description{
Extracts the keys from a \code{map} object.
}
