% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peek.R
\docType{methods}
\name{peek}
\alias{peek}
\alias{peek,heap-method}
\alias{peek,deque-method}
\title{Have a look at the first element from an object without removing it}
\usage{
peek(obj)

\S4method{peek}{heap}(obj)

\S4method{peek}{deque}(obj)
}
\arguments{
\item{obj}{the object to peek}
}
\value{
returns the first element from \code{obj}
}
\description{
Peeks into an object, i.e. takes the first element and returns
 it without removing it from the object.
}
