% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peek.R, R/ds_deque_queue.R,
%   R/ds_deque_stack.R, R/ds_heap.R
\docType{methods}
\name{peek}
\alias{peek}
\alias{peek,queue-method}
\alias{peek,stack-method}
\alias{peek,heap-method}
\title{Have a look at the first element from an object without removing it}
\usage{
peek(obj)

\S4method{peek}{queue}(obj)

\S4method{peek}{stack}(obj)

\S4method{peek}{heap}(obj)
}
\arguments{
\item{obj}{the object to peek}
}
\value{
returns the first element from \code{obj} as list
}
\description{
Peeks into an object, i.e. takes the first element and returns
 it without removing it from the object.
}
