% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peek.R, R/ds_fibonacci_heap.R,
%   R/ds_queue.R, R/ds_stack.R
\docType{methods}
\name{peek}
\alias{peek}
\alias{peek,fibonacci_heap-method}
\alias{peek,queue-method}
\alias{peek,stack-method}
\title{Have a look at the first element from an object without removing it}
\usage{
peek(obj)

\S4method{peek}{fibonacci_heap}(obj)

\S4method{peek}{queue}(obj)

\S4method{peek}{stack}(obj)
}
\arguments{
\item{obj}{the object to peek into}

\item{obj}{the object to peek into}

\item{obj}{the object to peek into}
}
\value{
returns the first element from the queue

returns the first element from the stack
}
\examples{
 fibonacci_heap <- new("fibonacci_heap", "character", "character")
 fibonacci_heap <- insert(fibonacci_heap, paste0("k", 1:10), paste0("v", 1:10))

 peek(hashmap)
 queue <- new("queue", "character")
 queue <- insert(queue, paste0("k", 1:10))

 peek(queue)
 stack <- new("stack", "character")
 stack <- insert(stack, paste0("k", 1:10))

 peek(stack)
}
