% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R, R/fibonacci_heap.R
\docType{methods}
\name{peek}
\alias{peek}
\alias{peek,fibonacci_heap-method}
\title{Have a look at the first element from an object without removing it}
\usage{
peek(obj)

\S4method{peek}{fibonacci_heap}(obj)
}
\arguments{
\item{obj}{the object to peek into}
}
\examples{
 fibonacci_heap <- new("fibonacci_heap", "character", "character")
 fibonacci_heap <- insert(hashmap, paste0("k", 1:10), paste0("v", 1:10))

 peek(hashmap)
}
