% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_pop.R, R/ds_fibonacci_heap.R,
%   R/ds_queue.R, R/ds_stack.R
\docType{methods}
\name{pop}
\alias{pop}
\alias{pop,fibonacci_heap-method}
\alias{pop,queue-method}
\alias{pop,stack-method}
\title{Pop a single element from an object}
\usage{
pop(obj)

\S4method{pop}{fibonacci_heap}(obj)

\S4method{pop}{queue}(obj)

\S4method{pop}{stack}(obj)
}
\arguments{
\item{obj}{the object to pop an element from}
}
\value{
returns the first element
}
\description{
Remove and return the first element from an object.
}
\examples{
 fibonacci_heap <- new("fibonacci_heap", "character", "character")
 fibonacci_heap <- insert(fibonacci_heap,
                          paste0("k", 1:10),
                          paste0("v", 1:10))

 pop(hashmap)
 queue <- new("queue", "character")
 queue <- insert(queue, paste0("k", 1:10))

 pop(queue)
 stack <- new("stack", "character")
 stack <- insert(stack, paste0("k", 1:10))

 pop(stack)
}
