% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_pop.R
\docType{methods}
\name{pop}
\alias{pop}
\alias{pop,deque-method}
\alias{pop,heap-method}
\title{Pop a single element from an object}
\usage{
pop(obj)

\S4method{pop}{deque}(obj)

\S4method{pop}{heap}(obj)
}
\arguments{
\item{obj}{the object to pop an element from}
}
\value{
returns the first element from \code{obj} as list
}
\description{
Remove and return the first element from an object.
}
