% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R, R/fibonacci_heap.R
\docType{methods}
\name{pop}
\alias{pop}
\alias{pop,fibonacci_heap-method}
\title{Pop a single element from an object}
\usage{
pop(obj)

\S4method{pop}{fibonacci_heap}(obj)
}
\arguments{
\item{obj}{the object to pop}
}
\examples{
 fibonacci_heap <- new("fibonacci_heap", "character", "character")
 fibonacci_heap <- insert(hashmap, paste0("k", 1:10), paste0("v", 1:10))

 pop(hashmap)
}
