% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_deque_queue.R
\docType{class}
\name{queue-class}
\alias{queue-class}
\title{Queue class}
\description{
Implementation of a queue data structure, i.e. a list
 implementation with FIFO principle. \code{queue} uses a \code{std::deque}
 as default container, so inserting, peeking and popping functions require
 constant \emph{O(1)}. See \code{\linkS4class{stack}} for a class using
 the LIFO principle.
}
\section{Slots}{

\describe{
\item{\code{.deque}}{\code{C++} object representing a deque}
}}

