% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_deque_queue.R
\name{queue}
\alias{queue}
\title{Create a new \code{queue}}
\usage{
queue()
}
\arguments{
\item{key.class}{the primitive class type of the keys}
}
\value{
returns a new \code{queue} object
}
\description{
Instantiates a new \code{\linkS4class{queue}} object,
 i.e. a list implementation with FIFO principle.
}
