% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_remove.R, R/ds_map.R, R/ds_map_bimap.R,
%   R/ds_map_multimap.R
\docType{methods}
\name{remove}
\alias{remove}
\alias{remove,map,vector,missing-method}
\alias{remove,bimap,missing,vector-method}
\alias{remove,multimap,vector,vector-method}
\alias{remove,multimap,vector,list-method}
\alias{remove,multimap,vector,ANY-method}
\title{Remove an entry from a map}
\usage{
remove(obj, key, value)

\S4method{remove}{map,vector,missing}(obj, key)

\S4method{remove}{bimap,missing,vector}(obj, value)

\S4method{remove}{multimap,vector,vector}(obj, key, value)

\S4method{remove}{multimap,vector,list}(obj, key, value)

\S4method{remove}{multimap,vector,ANY}(obj, key, value)
}
\arguments{
\item{obj}{the object to pop an element from}

\item{key}{a vector of keys that should be removed}

\item{value}{optionally a list of values needs to be supplied for some data
structures such as multimaps if entries should be removed.}
}
\value{
returns \code{obj} with removed values
}
\description{
Remove a vector of key-value pair from a map object.
}
