% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_size.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,heap-method}
\alias{size,map-method}
\alias{size,deque-method}
\title{Get the size of an object}
\usage{
size(obj)

\S4method{size}{heap}(obj)

\S4method{size}{map}(obj)

\S4method{size}{deque}(obj)
}
\arguments{
\item{obj}{the object to get the size from}
}
\value{
returns the size of an object
}
\description{
Computes the size of an object, i.e. the number of keys or
 <key, value> pairs depending on the object.
}
