% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_size.R, R/ds_deque.R, R/ds_heap.R,
%   R/ds_map.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,deque-method}
\alias{size,heap-method}
\alias{size,map-method}
\title{Get the size of an object}
\usage{
size(obj)

\S4method{size}{deque}(obj)

\S4method{size}{heap}(obj)

\S4method{size}{map}(obj)
}
\arguments{
\item{obj}{the object to get the size from}
}
\value{
returns the size of \code{obj}
}
\description{
Computes the size of an object, i.e. the number of keys or
 <key, value> pairs depending on the object.
}
