% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_size.R, R/ds_fibonacci_heap.R,
%   R/ds_hashmap.R, R/ds_queue.R, R/ds_stack.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,fibonacci_heap-method}
\alias{size,hashmap-method}
\alias{size,queue-method}
\alias{size,stack-method}
\title{Get the size of an object}
\usage{
size(obj)

\S4method{size}{fibonacci_heap}(obj)

\S4method{size}{hashmap}(obj)

\S4method{size}{queue}(obj)

\S4method{size}{stack}(obj)
}
\arguments{
\item{obj}{the object to get the size from}
}
\value{
returns the size of an object
}
\description{
Computes the size of an object, i.e. the number of keys or
 <key, value> pairs depending on the object.
}
