% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_deque_stack.R
\docType{class}
\name{stack-class}
\alias{stack-class}
\title{Stack class}
\description{
Implementation of a stack datastructure, i.e. a list
 implementation with LIFO principle. \code{stack} uses a \code{std::deque}
 as default container, so inserting, peeking and popping functions require
 constant \emph{O(1)}. See \code{\linkS4class{queue}} for a class using
 the FIFO principle
}
\section{Slots}{

\describe{
\item{\code{.deque}}{\code{C++} object representing a deque}

\item{\code{.key.class}}{the class of the keys}
}}

