% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_map_hashmap.R
\docType{methods}
\name{[,hashmap,ANY,missing,missing-method}
\alias{[,hashmap,ANY,missing,missing-method}
\title{Extract elements from an object}
\usage{
\S4method{[}{hashmap,ANY,missing,missing}(x, i)
}
\arguments{
\item{x}{a \code{hashmap}}

\item{i}{a vector of keys}
}
\description{
Access <key, value> pairs of a hashmap using a set of keys.
}
