% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_values.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,hashmap-method}
\alias{values,bimap-method}
\title{Get values from an object}
\usage{
values(obj)

\S4method{values}{hashmap}(obj)

\S4method{values}{bimap}(obj)
}
\arguments{
\item{obj}{object to extract values from}
}
\value{
returns the extracted values
}
\description{
Extracts the values from a \code{map} object.
}
